\name{CEMS-package}
\alias{CEMS-package}
\alias{CEMS}
\docType{package}
\title{
\packageTitle{CEMS}
}
\description{
\packageDescription{CEMS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CEMS}
\packageIndices{CEMS}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{CEMS}

Maintainer: \packageMaintainer{CEMS}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
